services/workflow-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0
openai==1.3.7
httpx==0.25.2
playwright==1.54.0
beautifulsoup4==4.13.4
lxml==6.0.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
scikit-learn==1.3.2
openpyxl==3.1.2
aiofiles==23.2.0
python-multipart==0.0.6
werkzeug==2.3.7

# Optional dependencies (install as needed)
# python-docx==0.8.11  # For Word document processing

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1